+++ cli/src/commands/extensions/api.rs
@@ -162,14 +162,14 @@ async fn analyze(
 
 /// Check a set of packages against the default policy.
 #[op]
-async fn package_check(
+async fn check_packages(
     op_state: Rc<RefCell<OpState>>,
     packages: Vec<PackageDescriptor>,
 ) -> Result<PolicyEvaluationResponse> {
     let state = ExtensionState::from(op_state);
     let api = state.api().await?;
 
-    Ok(api.package_check(&packages).await?)
+    Ok(api.check_packages(&packages).await?)
 }
 
 /// Retrieve user info.
@@ -484,7 +484,7 @@ async fn api_base_url(op_state: Rc<RefCell<OpState>>) -> Result<String> {
 pub(crate) fn api_decls() -> Vec<OpDecl> {
     vec![
         analyze::decl(),
-        package_check::decl(),
+        check_packages::decl(),
         get_user_info::decl(),
         get_access_token::decl(),
         get_refresh_token::decl(),