+++ extensions/phylum.ts
@@ -105,14 +105,14 @@ export class PhylumApi {
   }
 
   /**
-   * Check packages against the default policy.
+   * Check a list of packages against the default policy.
    *
    * @param packages - List of packages to check (see `analyze()` for details)
    *
    * @returns Job analysis results (see `getJobStatus()` for details)
    */
-  static packageCheck(packages: Package[]): Promise<Record<string, unknown>> {
-    return DenoCore.opAsync("package_check", packages);
+  static checkPackages(packages: Package[]): Promise<Record<string, unknown>> {
+    return DenoCore.opAsync("check_packages", packages);
   }
 
   /**